new test cases decl46 and 47 for multiple declarations of the same function or intrinsices; updated results, especially for optional 'extern' declarator
git-svn-id: http://svn.cri.ensmp.fr/svn/validation/trunk@1342 67b5017d-e62d-0410-a871-f72b1c62810f

+++ C_syntax/dependent_type01.c
@@ -1,3 +1,5 @@
+/* This code is C99. It is not parsed by the PIPS C parser. */
+
 #include <stdio.h>
 
 main()
@@ -8,6 +10,6 @@ main()
   m++;
   n++;
   double b[m][n];
-  printf("size of a = %d\n", sizeof(a));
-  printf("size of b = %d\n", sizeof(b));
+  printf("size of a = %d\n", (int) sizeof(a));
+  printf("size of b = %d\n", (int) sizeof(b));
 }
